pyproject.toml
[tool.poetry]
name = "sol"
version = "0.1.0"
description = "Sol - AI Trading Orchestrator for Indian Stock Market"
authors = ["Harsh"]
readme = "README.md"
packages = [{ include = "sol" }]

[tool.poetry.dependencies]
python = "^3.11"
kiteconnect = ">=4.2.0"
anthropic = ">=0.40.0"
openai = ">=1.0.0"
google-generativeai = ">=0.8.0"
fastapi = ">=0.115.0"
uvicorn = { version = ">=0.30.0", extras = ["standard"] }
sqlalchemy = { version = ">=2.0.0", extras = ["asyncio"] }
asyncpg = ">=0.30.0"
alembic = ">=1.13.0"
redis = { version = ">=5.0.0", extras = ["hiredis"] }
apscheduler = ">=3.10.0"
pydantic-settings = ">=2.0.0"
pandas = ">=2.0.0"
python-jose = { version = ">=3.3.0", extras = ["cryptography"] }
cryptography = ">=42.0.0"
httpx = ">=0.27.0"
structlog = ">=24.0.0"
python-multipart = ">=0.0.9"
passlib = { version = ">=1.7.4", extras = ["bcrypt"] }
pytz = ">=2024.1"
ta = ">=0.11.0"
aiohttp = ">=3.9.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.23.0"
pytest-cov = ">=5.0.0"
httpx = ">=0.27.0"
black = ">=24.0.0"
ruff = ">=0.4.0"
mypy = ">=1.10.0"
aiosqlite = ">=0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
